Movie_PrePlay: Fanart
Similar to the PrePlay page of PlexConnect - with a couple of issues:
- buttons (badge&text) are basically black when not highlighted: how to
modify “background color”?
- width/height styles don’t really seem to work - or I do something
totally wrong: how to get rid of the “img”s to set width/height of
neighbor elements?
- autoHighlight=“true” does not work in divTemplate (same in
decriptiveAlertTemplate): how to fix that - major drawback in
usability?
Anybody with better experience in CSS?

Utilities.swift: fall back to default theme if page is not available.
This limits the need for double-storing unchanged TVML stuff.